requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
requests==2.31.0
pypdf==5.5.0
rich==13.7.0
langchain>=0.1.0,<0.3.0
langchain-openai>=0.0.5,<0.2.0
langchain-community>=0.0.16,<0.3.0
langchain-text-splitters>=0.0.1,<0.3.0
langgraph>=0.0.20,<0.3.0
faiss-cpu==1.7.4
openai>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0
langgraph-checkpoint-mongodb>=0.1.0